Add environment classes and helpers
Get a setting from the environment through a helper.  This makes it possible to inject
values suitable for testing or maintenance commands without having to provide a comprehensive
set of values that do not usually apply to a test run or maintenance command.  For certain
maintenance commands however, some settings must be correct; like access to the database when
migrations are being applied.